+++ ydb/library/yql/providers/generic/provider/yql_generic_predicate_pushdown.cpp
@@ -341,6 +341,9 @@ namespace NYql {
             MATCH_ATOM(Utf8, UTF8, text, TString);
             MATCH_ATOM(Timestamp, TIMESTAMP, int64, i64);
             MATCH_ATOM(Interval, INTERVAL, int64, i64);
+            // Arrow vector holds value with ui16 type
+            // Proto value does not have ability to store
+            // ui16 type that's why uint32 is used
             MATCH_ATOM(Date, DATE, uint32, ui16);
             MATCH_ARITHMETICAL(Sub, SUB);
             MATCH_ARITHMETICAL(Add, ADD);
@@ -721,7 +724,7 @@ namespace NYql {
             }
             case Ydb::Type::DATE: {
                 const auto& value = typedValue.value();
-                switch(value.value_case()) {
+                switch (value.value_case()) {
                 case Ydb::Value::kUint32Value:
                     return TStringBuilder() << FormatType(typedValue.type()) << "(\""
                         << TInstant::Days(value.uint32_value()).FormatLocalTime("%Y-%m-%d") << "\")";
